begin refactor, breaks on ai ending turn
