+++ src/ai/ai_utils.rs
