--- src/ai_request.rs
@@ -1,4 +1,0 @@
-use rustc_serialize::json;
-use game_state::GameStateData;
-
-#[derive(RustcDecodable, RustcEncodable, Clone)]
